requirements.txt
openai
transformers
python-dotenv
fastapi
uvicorn[standard]
streamlit 
requests
wikipedia-api
colorama


